makeover

+++ EmailDomain.go
@@ -1,8 +1,6 @@
 package insightly
 
-// types
-//
 type EmailDomain struct {
-	EmailDomainID int    `json:"EMAIL_DOMAIN_ID"`
+	EmailDomainID int64  `json:"EMAIL_DOMAIN_ID"`
 	EmailDomain   string `json:"EMAIL_DOMAIN"`
 }
